version 1.1.0
Now the inscription "There was no shot in the previous move" is not displayed on the first move.

The use of color inversion has been reduced because older versions of Windows don't work well with them.

The welcome screen has been changed (when starting the game).

Now the game results screen has become more beautiful.